#1 Partial commit of some files necessary to get initial draft of website working
For the website, we plan to have links to some files for download.

Many of these are for the image perforation paper.

These files aren't small.

We're having trouble with making one large commit with EVERYTHING, so we're going to try and commit in parts.

This commits the image perforation files as well as some css files.