#1 Minor changes to README to reflect URL change
The previous name of the repository led to an unreasonably long URL, so we changed it to something simpler (i.e. https://paul-tqh-nguyen.github.io/about/ ) via changing the repository's name.

This patch changes the README to reflect this new name change and to add a link to the live demo.