#1 Change CSS to be more mobile friendly
Third party testers have noted that the webpage has aesthetic issus on mobile platforms.

This stems from the fact that the header-text block on teh "home" portion of the website is implemented as a button.

Block has a large width.

When this width exceed the size of a mobile device's width, horizontal scrolling happens because we cannot squish a button.

The header text block is now implemented in a way that uses DIV styling and not that of a button.

There still may be modile issues, but we'll address those as they arrive.