#1 Remove header row for mobile devices
This patch introduces a partial implementatino for the feature to remove the header row when we're displaying ona mobile device.

The intention behind this is to maximize the amount of real estate that useful content takes up on mobile devices.

The header row on mobile devices is mroe distracting than anything else.