#1 Work around display issue for unequal length Image Perforation example image descriptor text
The descriptor texts for Image Perforation example images aren't of the same length. When the text is displayed on different devices, it causes the relative place of other items (in particular the images) to be out of line.

This patch works around this by making the text more equal in length.