#1 Robustify mobile device classes for "home" portion of website
This patch fixes the specialized support we have for mobile devices where.

The implementation was previously kind of dirty in the sense that we toggled DOM elements with IDs to have a certain class when on mobile devices.

The problem was that different DOM elemenets had the same ID bc we wanted the same functionality.

This patch fixes the above problem my moving the styles over to classes instead of IDs.

Not a huge change, but makes for cleaner code.